Add android samples kotlin-lib, java-lib, kotlin-app
