+++ samples/android-app-java-sample/src/main/java/com/squareup/wire/android/app/java/MainActivity.java
@@ -13,10 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-syntax = "proto3";
+package com.squareup.wire.android.app.java;
 
-package com.squareup.wire.javaonly;
+import android.widget.TextView;
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
 
-message SomeText {
-  string value = 1;
+
+public class MainActivity extends AppCompatActivity {
+
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    setContentView(R.layout.activity_main);
+
+    SomeText someText = new SomeText("Hi");
+    ((TextView) findViewById(R.id.text_view)).setText(someText.value);
+  }
 }
