+++ samples/android-lib-kotlin-sample/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
-  id("com.android.application")
+  id("com.android.library")
+  id("org.jetbrains.kotlin.android")
   id("com.squareup.wire")
 }
 
@@ -8,11 +9,8 @@ android {
   buildToolsVersion("30.0.2")
 
   defaultConfig {
-    applicationId = "com.square.wire.java.sample.myapplication"
     minSdkVersion(28)
     targetSdkVersion(30)
-    versionCode = 1
-    versionName = "1.0"
   }
 
   compileOptions {
@@ -22,14 +20,10 @@ android {
 }
 
 wire {
-  java {
+  kotlin {
     android = true
+    javaInterop = true
   }
-}
-
-dependencies {
-  implementation("androidx.appcompat:appcompat:1.2.0")
-  implementation("androidx.constraintlayout:constraintlayout:2.0.4")
 }
 
 buildscript {
