+++ settings.gradle.kts
@@ -1,5 +1,8 @@
 include(":samples:simple-sample")
 include(":samples:android-app-java-sample")
+include(":samples:android-app-kotlin-sample")
+include(":samples:android-lib-java-sample")
+include(":samples:android-lib-kotlin-sample")
 include(":samples:wire-codegen-sample")
 include(":samples:wire-grpc-sample:client")
 include(":samples:wire-grpc-sample:protos")
